UPM Epassafe
UMP - Epassafe
--------------------------
http://www.epassafe.com

UMP - Epassafe (C) 2013 is a fork of Universal Password Manager (UMP)
with the
development geared to making this a stand-alone application on an
Android Device.

UMP - Epassafe differs from Universal Password Manager in several ways.
It has been
stripped down to the most basic usage, that of encrypting your passwords
and allowing
backups to be made to external device i.e sdcard. Also, this program
fork has been designed
to be the sole application to run aside from the most basic android
specific applications.

For example, Epassafe replaces the default Launcher.
I believe their is a niche market for requiring this kind of device.

Apart from how Epassafe differs from Universal Password Manager, the aim
is to use it on a
modified Android OS or special Rom to make the device sole purpose for
keeping passwords
safe, like a notebook but in electronic form.